requirements.txt
# BLUFF: Benchmark for Linguistic Understanding of Fake-news Forensics
# Python 3.10+ required

# Core
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
tokenizers>=0.15.0

# Multilingual NLP
sentencepiece>=0.1.99
protobuf>=3.20.0

# Evaluation
scikit-learn>=1.3.0
scipy>=1.11.0
seqeval>=1.2.2

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
tqdm>=4.66.0

# Language Detection
langdetect>=1.0.9
fasttext-wheel>=0.9.2

# API Clients (for decoder experiments)
openai>=1.6.0
anthropic>=0.8.0
google-generativeai>=0.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
pyyaml>=6.0.1
jsonlines>=4.0.0
python-dotenv>=1.0.0
huggingface-hub>=0.20.0

# Quality Filtering (mPURIFY)
sentence-transformers>=2.2.2
rouge-score>=0.1.2
nltk>=3.8.1
